Выделил логику оптимизирования нагрузки в отдельный класс
