+++ src/main/java/faang/school/godbless/javaHashMap/dataCenter/Main.java
@@ -45,7 +45,7 @@ public class Main {
 
         System.out.println("------------------------------------------");
 
-        dataCenterService.optimize(dataCenterService.getDataCenter());
+        dataCenterService.optimizeDataCenter();
 
         System.out.println("------------------------------------------");
